# cmd/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Meetin APP API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/meeting/Appointment/multiple": {
            "put": {
                "description": "Update Multiple in UpdateAppointment",
                "tags": [
                    "meeting Appointment"
                ],
                "summary": "Update Multiple in Appointment",
                "operationId": "ApiV1MeetingAppAppointmentsUpdateMultiple",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.AppointmentEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Appointment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create Multiple in CreateAppointment",
                "tags": [
                    "meeting Appointment"
                ],
                "summary": "Create Multiple in Appointment",
                "operationId": "ApiV1MeetingAppAppointmentCreateMultiple",
                "parameters": [
                    {
                        "type": "string",
                        "default": "[[authToken]]",
                        "description": "Current Session Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "[[companyID]]",
                        "description": "Current CompanyID",
                        "name": "x-company",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.AppointmentFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Appointment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/Appointment/pagination": {
            "get": {
                "description": "Pagination GetAppointmentWithPagination",
                "tags": [
                    "meeting Appointment"
                ],
                "summary": "Pagination Appointment",
                "operationId": "ApiV1MeetingAppAppointmentPagination",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Appointment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/Appointment/search": {
            "get": {
                "description": "Get GetAppointment",
                "tags": [
                    "meeting Appointment"
                ],
                "summary": "Show Appointment",
                "operationId": "ApiV1MeetingAppAppointmentSearch",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Appointment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/Appointment/single": {
            "post": {
                "description": "Create in CreateAppointment",
                "tags": [
                    "meeting Appointment"
                ],
                "summary": "Create in Appointment",
                "operationId": "ApiV1MeetingAppAppointmentCreate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.AppointmentFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_structures.Appointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/Appointment/with-id/:id": {
            "get": {
                "description": "Get GetAppointmentById",
                "tags": [
                    "meeting Appointment"
                ],
                "summary": "Get Appointment by Unique IDs",
                "operationId": "ApiV1MeetingAppAppointmentWithID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Appointment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update with id in UpdateAppointment",
                "tags": [
                    "meeting Appointment"
                ],
                "summary": "Update with id in Appointment",
                "operationId": "ApiV1MeetingAppAppointmentUpdateWithID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.AppointmentEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_structures.Appointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete with path params in DeleteAppointment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meeting Appointment"
                ],
                "summary": "Delete with path params in Appointment",
                "operationId": "ApiV1MeetingAppAppointmentDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_structures.Appointment": {
            "type": "object",
            "properties": {
                "appointmentDate": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "customerID": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "purpose": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        },
        "api_structures.AppointmentEdit": {
            "type": "object",
            "properties": {
                "appointmentDate": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "customerID": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "purpose": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        },
        "api_structures.AppointmentFilter": {
            "type": "object",
            "properties": {
                "appointmentDate": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "customerID": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "purpose": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        },
...
    }
}